Fix collection utils refactor imports
