+++ src/benchmark/framework/parseDataset/cacheStructures/cacheHolders.ts
@@ -1,13 +1,13 @@
 import { ProofGoal } from "../../../../coqLsp/coqLspTypes";
 
+import { all } from "../../../../utils/collectionUtils/listUtils";
+import { mapValues } from "../../../../utils/collectionUtils/mapUtils";
 import { joinPaths } from "../../../../utils/fs/pathUtils";
 import { CodeElementRange } from "../../../../utils/structures/codeElementPositions";
 import { TargetType } from "../../structures/benchmarkingCore/completionGenerationTask";
 import { TargetRequestType } from "../../structures/common/inputTargets";
 import { ParsedCoqFileData } from "../../structures/parsedCoqFile/parsedCoqFileData";
 import { TheoremData } from "../../structures/parsedCoqFile/theoremData";
-import { all } from "../../utils/collectionUtils/listUtils";
-import { mapValues } from "../../utils/collectionUtils/mapUtils";
 import { toTargetType } from "../../utils/commonStructuresUtils/targetTypeUtils";
 
 export class DatasetCacheHolder {
